requirements.txt
PyQt5==5.15.9
Pillow==9.5.0
pytesseract==0.3.10
pyperclip3==0.4.1
numpy==1.24.3
opencv-python==4.7.0.72